Add functions that used to get auther data like exchanges informations
