+++ README.md
@@ -23,3 +23,5 @@ How to call the cryptocompare.com API?
 `
 
 ___
+
+### Useful
